tools/requirements.txt
python-dotenv==1.0.0
oracledb==2.0.1
psycopg[binary]==3.1.16
pymongo==4.6.0
